chore: Add additional configs (#1)
This pull request includes several significant changes to the project,
focusing on the implementation of a new main program and the removal of
outdated example and test files. The most important changes include the
addition of a new `Makefile` for building and running the project, the
implementation of a new `main.go` file, and the removal of old example
and test files.

### New Implementations:

*
[`Makefile`](diffhunk://#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52R1-R13):
Added a new `Makefile` to streamline the build and run process. This
includes targets for cleaning up generated files and running the main
program.
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R1-R65):
Implemented a new main program that initializes and runs multiple
instances of the DIMEX module. This includes a worker function to
simulate the flow of an application using the DIMEX module.

### Removed Files:

*
[`testDIMEX.go`](diffhunk://#diff-81562b6a7733d62b4eb8fa431268574a0c2d9888a9fbbd7823686e984cc2cc02L1-L56):
Removed the old test file that checked the correctness of the output
file generated by the DIMEX module.
*
[`useDIMEX-f.go`](diffhunk://#diff-5a9fc7a5038879772b10cf48dc5ca90fde0f1a80c8f20cff59e280e81cbba4e8L1-L94):
Removed the old example file that demonstrated the use of the DIMEX
module in a distributed system.