fix: Algorithm for taking snapshots (#5)
This pull request includes a change to the `DIMEX/DIMEX-Template.go`
file to modify the behavior of the `Start` function within the
`DIMEX_Module` struct. The most important change involves altering how
snapshots are initiated by sending a snapshot message to the module
itself instead of directly calling the `takeSnapshot` method. This
effectively fixes the issue where the snapshots were taken
inconsistently, as the process that initiates the snapshot was taking
its own snapshot concurrently with the other module actions related to
handling the DiMEx messages.

Changes to snapshot handling:

*
[`DIMEX/DIMEX-Template.go`](diffhunk://#diff-21e10de124c5cf8ae4d7061f9e6ef652bb847a15ebfe6863812183f67d308902L157-R162):
In the `Start` function, replaced the direct call to
`module.takeSnapshot(snapId)` with a call to `module.sendToLink`,
sending a formatted snapshot message to the module's own address.