fix: Ensure proper SNAP resps counting and closure for communication channels for messages in transit (#11)
This pull request introduces significant updates to the snapshot
functionality in the `Dimex` system, focusing on improving the handling
of messages in transit and ensuring that the communication channels
between processes are properly closed, as well as properly counting the
number of received SNAP responses to consider the snapshot finished and
then dump it.

### Snapshot Handling Enhancements:
* Replaced the `InterceptedMsgs` field in the `Snapshot` struct with a
`CommunicationChans` map to track messages in transit and the status of
communication channels. (`snapshots/snapshots.go`,
[snapshots/snapshots.goL25-R80](diffhunk://#diff-7445aedbbb0091d50bab67428e19d87799abdc13295ec57d756fbe731dc93bf8L25-R80))
* Introduced a `NewSnapshot` constructor to initialize snapshots
structs, including setting up communication channels for all processes.
(`snapshots/snapshots.go`, F63b95f7L43R102)
* Added a `HasMessagesInTransit` method to the `Snapshot` struct to
check if there are any messages in transit. (`snapshots/snapshots.go`,
[snapshots/snapshots.goR105-R114](diffhunk://#diff-7445aedbbb0091d50bab67428e19d87799abdc13295ec57d756fbe731dc93bf8R105-R114))

### Updates to Snapshot Logic:
* Updated the `takeSnapshot` method in `dimex.go` to use the new
`NewSnapshot` constructor and include the sender's process ID in
snapshot messages. (`dimex/dimex.go`,
[dimex/dimex.goL360-R411](diffhunk://#diff-f0b76059466e719392bdd9a874e042cd383c90990850193feb14dffab2e79973L360-R411))
* Modified the `handleIncomingSnap` method to account for the sender ID
when processing snapshot responses and updating communication channel
states. (`dimex/dimex.go`,
[dimex/dimex.goL310-R336](diffhunk://#diff-f0b76059466e719392bdd9a874e042cd383c90990850193feb14dffab2e79973L310-R336))
* Don't consider the SNAP request sent from a process to itself (when it
takes the initiative of starting a snapshot) as a response in the sense
of incrementing the `CollectedResps` field of the snapshot. This was
causing a silent bug where snapshot initiators were finishing the
snapshot when a SNAP resp was still missing, because it was considering
the SNAP request received from itself as a SNAP response already.

### Invariant Checks Adjustments:
* Updated the `checkIdleProcessesState` function to use the new
`HasMessagesInTransit` method instead of directly checking the
`InterceptedMsgs` field. (`snapshots/invariants.go`,
[snapshots/invariants.goL83-R84](diffhunk://#diff-9e3fbbc65495d162084dd1383f58aa95d0c49be7330ea55d4a08b4e7aef010b9L83-R84))
* Adjusted comments in `checkIdleProcessesState` to reflect the new
terminology ("messages in transit" instead of "intercepted messages").
(`snapshots/invariants.go`,
[snapshots/invariants.goL60-R68](diffhunk://#diff-9e3fbbc65495d162084dd1383f58aa95d0c49be7330ea55d4a08b4e7aef010b9L60-R68))